+++ src/rpucuda/cuda/rpucuda_transfer_device.h
@@ -52,6 +52,8 @@ public:
   };
   TransferRPUDeviceCuda<T> *clone() const override { return new TransferRPUDeviceCuda<T>(*this); };
 
+  void setHiddenUpdateIdx(int idx) override{};
+
   void runUpdateKernel(
       pwukp_t<T> kpars,
       CudaContext *up_context,